Move AdView and MapView to package directories
