+++ core/ui/src/androidMain/kotlin/org/pointyware/xyz/ui/ads/AdView.android.kt
@@ -2,7 +2,7 @@
  * Copyright (c) 2024 Pointyware. Use of this software is governed by the GPL-3.0 license.
  */
 
-package org.pointyware.xyz.core.ui
+package org.pointyware.xyz.ui.ads
 
 import android.annotation.SuppressLint
 import androidx.compose.runtime.Composable
