+++ core/ui/src/androidMain/kotlin/org/pointyware/xyz/ui/maps/MapView.android.kt
@@ -1,4 +1,4 @@
-package org.pointyware.xyz.core.ui.maps
+package org.pointyware.xyz.ui.maps
 
 import androidx.compose.runtime.Composable
 import androidx.fragment.compose.AndroidFragment
