+++ core/ui/src/commonMain/kotlin/org/pointyware/xyz/ui/ads/AdView.kt
@@ -2,21 +2,22 @@
  * Copyright (c) 2024 Pointyware. Use of this software is governed by the GPL-3.0 license.
  */
 
-package org.pointyware.xyz.core.ui
+package org.pointyware.xyz.ui.ads
 
 import androidx.compose.runtime.Composable
 import androidx.compose.ui.Modifier
 
+data class AdViewState(
+    val unit: String,
+)
 
-actual fun testAdViewState(): AdViewState = TODO()
+expect fun testAdViewState(): AdViewState
 
 /**
- *
+ * Displays an ad.
  */
 @Composable
-actual fun AdView(
+expect fun AdView(
     state: AdViewState,
-    modifier: Modifier
-) {
-
-}
+    modifier: Modifier = Modifier
+)
