+++ core/ui/src/jvmMain/kotlin/org/pointyware/xyz/ui/ads/AdView.jvm.kt
@@ -2,7 +2,7 @@
  * Copyright (c) 2024 Pointyware. Use of this software is governed by the GPL-3.0 license.
  */
 
-package org.pointyware.xyz.core.ui
+package org.pointyware.xyz.ui.ads
 
 import androidx.compose.runtime.Composable
 import androidx.compose.ui.Modifier
